chore: replace legacy `aggregate` with `AggregateOperation`
