+++ lib/collection.js
@@ -35,7 +35,7 @@ const removeDocuments = require('./operations/collection_ops').removeDocuments;
 const save = require('./operations/collection_ops').save;
 const updateDocuments = require('./operations/collection_ops').updateDocuments;
 
-const AggregateOperation = require('./operations/aggregate_operation');
+const AggregateOperation = require('./operations/aggregate');
 const BulkWriteOperation = require('./operations/bulk_write');
 const CountDocumentsOperation = require('./operations/count_documents');
 const CreateIndexOperation = require('./operations/create_index');
