+++ lib/db.js
@@ -32,7 +32,7 @@ const listCollectionsTransforms = require('./operations/db_ops').listCollections
 const profilingInfo = require('./operations/db_ops').profilingInfo;
 const validateDatabaseName = require('./operations/db_ops').validateDatabaseName;
 
-const AggregateOperation = require('./operations/aggregate_operation');
+const AggregateOperation = require('./operations/aggregate');
 const AddUserOperation = require('./operations/add_user');
 const CollectionsOperation = require('./operations/collections');
 const CommandOperation = require('./operations/command');
