+++ lib/operations/aggregate.js
@@ -1,132 +1,122 @@
 'use strict';
 
-const AggregationCursor = require('../aggregation_cursor');
-const applyWriteConcern = require('../utils').applyWriteConcern;
-const decorateWithCollation = require('../utils').decorateWithCollation;
-const decorateWithReadConcern = require('../utils').decorateWithReadConcern;
-const handleCallback = require('../utils').handleCallback;
+const CommandOperationV2 = require('./command_v2');
 const MongoError = require('../core').MongoError;
-const resolveReadPreference = require('../utils').resolveReadPreference;
-const toError = require('../utils').toError;
+const maxWireVersion = require('../core/utils').maxWireVersion;
 const ReadPreference = require('../core').ReadPreference;
+const Aspect = require('./operation').Aspect;
+const defineAspects = require('./operation').defineAspects;
 
 const DB_AGGREGATE_COLLECTION = 1;
 const MIN_WIRE_VERSION_$OUT_READ_CONCERN_SUPPORT = 8;
 
-/**
- * Perform an aggregate operation. See Collection.prototype.aggregate or Db.prototype.aggregate for more information.
- *
- * @method
- * @param {Db} db A Db instance.
- * @param {Collection|string} coll A collection instance or the string '1', used for db.aggregate.
- * @param {object} [pipeline=[]] Array containing all the aggregation framework commands for the execution.
- * @param {object} [options] Optional settings. See Collection.prototype.aggregate or Db.prototype.aggregate for a list of options.
- * @param {Db~aggregationCallback|Collection~aggregationCallback} callback The command result callback
- */
-function aggregate(db, coll, pipeline, options, callback) {
-  const isDbAggregate = typeof coll === 'string';
-  const target = isDbAggregate ? db : coll;
-  const topology = target.s.topology;
-  let hasWriteStage = false;
+class AggregateOperation extends CommandOperationV2 {
+  constructor(parent, pipeline, options) {
+    // ensure we receive an unchanged raw response from the server (for cursor logic)
+    options.full = true;
+    super(parent, options);
 
-  if (typeof options.out === 'string') {
-    pipeline = pipeline.concat({ $out: options.out });
-    hasWriteStage = true;
-  } else if (pipeline.length > 0) {
-    const finalStage = pipeline[pipeline.length - 1];
-    if (finalStage.$out || finalStage.$merge) {
-      hasWriteStage = true;
+    this.target =
+      parent.s.namespace && parent.s.namespace.collection
+        ? parent.s.namespace.collection
+        : DB_AGGREGATE_COLLECTION;
+
+    this.pipeline = pipeline;
+
+    // determine if we have a write stage, override read preference if so
+    this.hasWriteStage = false;
+    if (typeof options.out === 'string') {
+      this.pipeline = this.pipeline.concat({ $out: options.out });
+      this.hasWriteStage = true;
+    } else if (pipeline.length > 0) {
+      const finalStage = pipeline[pipeline.length - 1];
+      if (finalStage.$out || finalStage.$merge) {
+        this.hasWriteStage = true;
+      }
+    }
+
+    if (this.hasWriteStage) {
+      this.readPreference = ReadPreference.primary;
     }
   }
 
-  let command;
-  let namespace;
-  let optionSources;
+  execute(server, callback) {
+    const options = this.options;
+    const serverWireVersion = maxWireVersion(server);
+    const inTransaction = this.session && this.session.inTransaction();
 
-  if (isDbAggregate) {
-    command = { aggregate: DB_AGGREGATE_COLLECTION, pipeline: pipeline };
-    namespace = db.s.namespace.withCollection(DB_AGGREGATE_COLLECTION);
+    const command = { aggregate: this.target, pipeline: this.pipeline };
 
-    optionSources = { db };
-  } else {
-    command = { aggregate: coll.collectionName, pipeline: pipeline };
-    namespace = coll.s.namespace;
-
-    optionSources = { db: coll.s.db, collection: coll };
-  }
-
-  const takesWriteConcern = topology.capabilities().commandsTakeWriteConcern;
-  const ismaster = topology.lastIsMaster() || {};
-
-  if (!hasWriteStage || ismaster.maxWireVersion >= MIN_WIRE_VERSION_$OUT_READ_CONCERN_SUPPORT) {
-    decorateWithReadConcern(command, target, options);
-  }
-
-  if (hasWriteStage && takesWriteConcern) {
-    applyWriteConcern(command, optionSources, options);
-  }
-
-  try {
-    decorateWithCollation(command, target, options);
-  } catch (err) {
-    if (typeof callback === 'function') return callback(err, null);
-    throw err;
-  }
-
-  if (options.bypassDocumentValidation === true) {
-    command.bypassDocumentValidation = options.bypassDocumentValidation;
-  }
-
-  if (typeof options.allowDiskUse === 'boolean') command.allowDiskUse = options.allowDiskUse;
-  if (typeof options.maxTimeMS === 'number') command.maxTimeMS = options.maxTimeMS;
-
-  if (options.hint) command.hint = options.hint;
-
-  options = Object.assign({}, options);
-
-  // Ensure we have the right read preference inheritance
-  options.readPreference = hasWriteStage
-    ? ReadPreference.primary
-    : resolveReadPreference(isDbAggregate ? db : coll, options);
-
-  if (options.explain) {
-    if (command.readConcern || command.writeConcern) {
-      throw toError('"explain" cannot be used on an aggregate call with readConcern/writeConcern');
-    }
-    command.explain = options.explain;
-  }
-
-  if (typeof options.comment === 'string') command.comment = options.comment;
-
-  // Validate that cursor options is valid
-  if (options.cursor != null && typeof options.cursor !== 'object') {
-    throw toError('cursor options must be an object');
-  }
-
-  options.cursor = options.cursor || {};
-  if (options.batchSize && !hasWriteStage) {
-    options.cursor.batchSize = options.batchSize;
-  }
-
-  command.cursor = options.cursor;
-
-  // promiseLibrary
-  options.promiseLibrary = target.s.promiseLibrary;
-
-  // Set the AggregationCursor constructor
-  options.cursorFactory = AggregationCursor;
-
-  if (typeof callback !== 'function') {
-    if (!topology.capabilities()) {
-      throw new MongoError('cannot connect to server');
+    if (
+      (!this.hasWriteStage || serverWireVersion >= MIN_WIRE_VERSION_$OUT_READ_CONCERN_SUPPORT) &&
+      !inTransaction
+    ) {
+      Object.assign(command, { readConcern: this.readConcern });
     }
 
-    return topology.cursor(namespace.toString(), command, options);
-  }
+    if (serverWireVersion >= 5) {
+      if (this.hasWriteStage && this.writeConcern) {
+        Object.assign(command, { writeConcern: this.writeConcern });
+      }
 
-  return handleCallback(callback, null, topology.cursor(namespace.toString(), command, options));
+      if (options.collation && typeof options.collation === 'object') {
+        Object.assign(command, { collation: options.collation });
+      }
+    }
+
+    if (options.bypassDocumentValidation === true) {
+      command.bypassDocumentValidation = options.bypassDocumentValidation;
+    }
+
+    if (typeof options.allowDiskUse === 'boolean') {
+      command.allowDiskUse = options.allowDiskUse;
+    }
+
+    if (typeof options.maxTimeMS === 'number') {
+      command.maxTimeMS = options.maxTimeMS;
+    }
+
+    if (options.hint) {
+      command.hint = options.hint;
+    }
+
+    if (options.explain) {
+      if (command.readConcern || command.writeConcern) {
+        callback(
+          new MongoError(
+            '"explain" cannot be used on an aggregate call with readConcern/writeConcern'
+          )
+        );
+
+        return;
+      }
+
+      command.explain = options.explain;
+    }
+
+    if (typeof options.comment === 'string') {
+      command.comment = options.comment;
+    }
+
+    // Validate that cursor options is valid
+    if (options.cursor != null && typeof options.cursor !== 'object') {
+      callback(new MongoError('cursor options must be an object'));
+      return;
+    }
+
+    command.cursor = options.cursor || {};
+    if (options.batchSize && !this.hasWriteStage) {
+      command.cursor.batchSize = options.batchSize;
+    }
+
+    super.executeCommand(server, command, callback);
+  }
 }
 
-module.exports = {
-  aggregate
-};
+defineAspects(AggregateOperation, [
+  Aspect.READ_OPERATION,
+  Aspect.RETRYABLE,
+  Aspect.EXECUTE_WITH_SELECTION
+]);
+
+module.exports = AggregateOperation;
