+++ lib/operations/count_documents.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const AggregateOperation = require('./aggregate_operation');
+const AggregateOperation = require('./aggregate');
 
 class CountDocumentsOperation extends AggregateOperation {
   constructor(collection, query, options) {
